Moving PackageRepository.Default into its own EnsoPackageRepository
