+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/EnsoContext.java
@@ -156,7 +156,7 @@ public class EnsoContext {
     var resourceManager = new org.enso.distribution.locking.ResourceManager(lockManager);
 
     packageRepository =
-        PackageRepository.initializeRepository(
+        DefaultPackageRepository.initializeRepository(
             OptionConverters.toScala(projectPackage),
             OptionConverters.toScala(languageHome),
             OptionConverters.toScala(editionOverride),
